# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News Aggregator API",
        "description": "News Aggregator API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://localhost:8000",
            "description": "Development Server"
        }
    ],
    "paths": {
        "/api/articles": {
            "get": {
                "tags": [
                    "Article"
                ],
                "summary": "Get articles",
                "description": "Get paginated list of articles",
                "operationId": "7e370449b00151b6422003b99ee09b88",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "path",
                        "description": "Search keyword within articles' title or description",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "debate"
                        }
                    },
                    {
                        "name": "source_id",
                        "in": "path",
                        "description": "Article by source",
                        "required": false,
                        "schema": {
                            "type": "int",
                            "example": "1"
                        }
                    },
                    {
                        "name": "author_id",
                        "in": "path",
                        "description": "Article by author",
                        "required": false,
                        "schema": {
                            "type": "int",
                            "example": "2"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "path",
                        "description": "Article by category",
                        "required": false,
                        "schema": {
                            "type": "int",
                            "example": "3"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of articles retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Sample Article Title"
                                            },
                                            "content": {
                                                "type": "string",
                                                "example": "This is the content of the article."
                                            },
                                            "author": {
                                                "type": "string",
                                                "example": "John Doe"
                                            },
                                            "published_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2023-01-01T12:00:00Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Authentication required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/articles/{id}": {
            "get": {
                "tags": [
                    "Article"
                ],
                "summary": "Get article by ID",
                "description": "Retrieve detailed information for a specific article by providing the article ID.",
                "operationId": "getArticleById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the article to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Sample Article Title"
                                        },
                                        "content": {
                                            "type": "string",
                                            "example": "This is the content of the article."
                                        },
                                        "author": {
                                            "type": "string",
                                            "example": "John Doe"
                                        },
                                        "published_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-01-01T12:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Article not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Authentication required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}